requirements.txt
amqp==5.3.1
apturl==0.5.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
billiard==4.2.1
black==24.8.0
blinker==1.4
Brlapi==0.7.0
cattrs==24.1.3
celery==5.5.1
certifi==2019.11.28
chardet==3.0.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.3
command-not-found==0.3
coverage==7.6.1
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.20
django-allauth==65.8.0
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docstring-to-markdown==0.17
duplicity==0.8.12.0
entrypoints==0.3
exceptiongroup==1.3.0
fasteners==0.14.1
future==0.18.2
greenlet==0.4.15
httplib2==0.14.0
idna==2.8
importlib-metadata==8.5.0
jedi==0.19.2
jedi-language-server==0.42.0
keyring==18.0.1
kombu==5.5.3
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lsprotocol==2023.0.1
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
msgpack==0.6.2
mypy-extensions==1.0.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==24.2
paramiko==2.6.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.6.0
Pillow==7.0.0
platformdirs==4.3.6
prompt-toolkit==3.0.51
protobuf==3.6.1
psycopg==3.2.6
pycairo==1.16.2
pycups==1.9.73
pygls==1.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvim==0.4.1
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
redis==5.2.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
sentry-sdk==2.28.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.5.3
systemd-python==234
tomli==2.2.1
typing-extensions==4.12.2
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==2.2.3
usb-creator==0.3.7
vine==5.1.0
wadllib==1.3.3
wcwidth==0.2.13
xkit==0.0.0
zipp==3.20.2
